+++ src/index.ts
@@ -3,5 +3,6 @@
 export { default as detectFocusLeft } from "./helpers/detectFocusLeft";
 export { default as focusFirst } from "./helpers/focusFirst";
 export { default as nestedProperty } from "./helpers/nestedProperty";
+export { default as onEvent } from "./helpers/onEvent";
 export { default as promiseWait } from "./helpers/promiseWait";
 export { default as stringPrettify } from "./helpers/stringPrettify";