+++ test/jest/helper.detectFocusLeft.test.js
@@ -1,6 +1,3 @@
-/**
- * @jest-environment jsdom
- */
 import detectFocusLeft from "web-ui-pack/helpers/detectFocusLeft";
 
 // use legacy accprding to https://stackoverflow.com/questions/68016538/matcher-error-received-value-must-be-a-mock-or-spy-function