+++ test/jest/helper.focusFirst.test.js
@@ -1,6 +1,3 @@
-/**
- * @jest-environment jsdom
- */
 import focusFirst from "web-ui-pack/helpers/focusFirst";
 
 afterEach(() => {