+++ src/index.ts
@@ -1,6 +1,5 @@
 // export * from "./popupElement";
 
-export { default as detectFocusLeft } from "./helpers/detectFocusLeft";
 export { default as focusFirst } from "./helpers/focusFirst";
 export { default as nestedProperty } from "./helpers/nestedProperty";
 export { default as onEvent } from "./helpers/onEvent";