+++ src/index.ts
@@ -3,6 +3,7 @@
 export { default as focusFirst } from "./helpers/focusFirst";
 export { default as nestedProperty } from "./helpers/nestedProperty";
 export { default as onEvent } from "./helpers/onEvent";
+export { default as onFocusGot } from "./helpers/onFocusGot";
 export { default as onFocusLost } from "./helpers/onFocusLost";
 export { default as promiseWait } from "./helpers/promiseWait";
 export { default as stringPrettify } from "./helpers/stringPrettify";