+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added [helpers](https://github.com/Yegorich555/web-ui-pack#helpers)
   - `stringPrettify`
   - `onEvent`
+  - `onFocusLost`
 
 ## 0.0.2 (Nov 30, 2021)