+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - `onEvent`
   - `onFocusGot`
   - `onFocusLost`
+  - `observer`
 - Removed helper `detectFocusLeft` in favor `onFocusLost`
 
 ## 0.0.2 (Nov 30, 2021)